refactor(extension): CLICK handler useRealMouse 抽 cdpClickElement #0000
dom.ts CLICK handler L104-210（useRealMouse 分支，107 行）抽出为
adapter/cdp.ts cdpClickElement(debuggerMgr, tabId, frameId, selector)。
CLICK handler 壳保留，分支改 1 行调用。

cdpClickElement 内部组合 native pageQuery（actionability 探测 + 取坐标，
page-side func 字面量保留在函数体内不跨文件抽，遵循 chrome.scripting
序列化限制）+ clickBBox（CDP attach + dispatchMouseEvent×3 trusted）。

scripting world: MAIN（page-side 探测）。debuggerMgr 显式参数。

dom.ts 1417 → 1313（-104），cdp.ts 33 → 151（+118）。
build ok，test 6 failed / 154 passed（基线，0 回归）。
